requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
fastapi==0.104.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
jose==1.0.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
pymongo
motor

